refactor(cli): move onboarding gateway and ollama helpers
